+++ frontend/src/api/contracts.api.ts
@@ -12,7 +12,7 @@ export const contractsApi: ContractsApi = {
     }).then(response => response.data),
     addContract: (newContract: NewContract) => axiosInstance.post('/contracts', newContract),
     updateContract: (contractId: number, contract: NewContract) => axiosInstance.put(`/contracts/${contractId}`, contract),
-    deleteContract: (contractId: number, details: DeleteContractDTO) => axiosInstance.delete(`/contracts/${contractId}`, {data: details}),
+    deleteContract: (details: DeleteContractDTO) => axiosInstance.delete(`/contracts`, {data: details}),
     generateMouthPayments: (yearMonth: string) => axiosInstance.post(`/contracts/generateMonthlyPayments/${yearMonth}`)
 }
 
@@ -22,6 +22,6 @@ export type ContractsApi = {
     fetchUnassignedRooms: (startDate: string, endDate: string) => Promise<Room[]>,
     addContract: (newContract: NewContract) => Promise<void>
     updateContract: (contractId: number, contract: NewContract) => Promise<void>
-    deleteContract: (contractId: number, details: DeleteContractDTO) => Promise<void>
+    deleteContract: (details: DeleteContractDTO) => Promise<void>
     generateMouthPayments: (yearMonth: string) => Promise<void>
 };
